+++ packages/dart_firebase_admin/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `FirebaseAdminApp.close`, to close open connections and stop the SDK.
 - Fixed various typos
+- Added `Firestore.collectionGroup` support
 
 ## 0.1.0 - 2023-10-15